# project/public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/sos": {
      "post": {
        "description": "Create a new SOS emergency request (PATIENT or EMERGENCY_CENTER)",
        "operationId": "SosController_createEmergencyRequest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmergencyRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Emergency request created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create emergency request",
        "tags": [
          "SOS / Emergency Requests"
        ]
      },
      "get": {
        "description": "Get all emergency requests created by the authenticated patient",
        "operationId": "SosController_getEmergencyRequests",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Emergency requests retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - PATIENT role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get my emergency requests",
        "tags": [
          "SOS / Emergency Requests"
        ]
      }
    },
    "/sos/{id}/assign": {
      "post": {
        "description": "Assign an emergency request to a specific hospital (EMERGENCY_CENTER only)",
        "operationId": "SosController_assignToHospital",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Emergency request ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "hospitalId": {
                    "type": "string",
                    "description": "Hospital organization ID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Emergency assigned to hospital successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - EMERGENCY_CENTER role required"
          },
          "404": {
            "description": "Emergency request or hospital not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign emergency to hospital",
        "tags": [
          "SOS / Emergency Requests"
        ]
      }
    },
    "/sos/{id}/status": {
      "put": {
        "description": "Update the status of an emergency request",
        "operationId": "SosController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Emergency request ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmergencyStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Emergency status updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Emergency request not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update emergency status",
        "tags": [
          "SOS / Emergency Requests"
        ]
      }
    },
    "/sos/all": {
      "get": {
        "description": "Get all emergency requests in the system (staff only)",
        "operationId": "SosController_getAllEmergencyRequests",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All emergency requests retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Staff role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all emergency requests",
        "tags": [
          "SOS / Emergency Requests"
        ]
      }
    },
    "/sos/dashboard/active-emergencies": {
      "get": {
        "description": "Get active emergencies (filtered by hospital if HOSPITAL role)",
        "operationId": "SosController_getActiveEmergencies",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Active emergencies retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Staff role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get active emergencies for dashboard",
        "tags": [
          "SOS / Emergency Requests"
        ]
      }
    },
    "/sos/{id}": {
      "get": {
        "description": "Get a specific emergency request by ID (patient only)",
        "operationId": "SosController_getEmergencyRequestById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Emergency request ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Emergency request retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - PATIENT role required"
          },
          "404": {
            "description": "Emergency request not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get emergency request by ID",
        "tags": [
          "SOS / Emergency Requests"
        ]
      }
    },
...
    "/api/settings/me": {
      "get": {
        "description": "Retrieve all settings for the authenticated user",
        "operationId": "SettingsController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User settings retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all user settings",
        "tags": [
          "Settings"
        ]
      },
      "put": {
        "description": "Update all or partial user settings (merges with existing)",
        "operationId": "SettingsController_putMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Settings updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user settings",
        "tags": [
          "Settings"
        ]
      }
    },
    "/api/settings/me/{category}": {
      "patch": {
        "description": "Update a specific category (notification, system, communication, profile, emergency)",
        "operationId": "SettingsController_patchCategory",
        "parameters": [
          {
            "name": "category",
            "required": true,
            "in": "path",
            "description": "Settings category (notification, system, communication, profile, emergency)",
            "schema": {
              "example": "notification",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category settings updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update specific settings category",
        "tags": [
          "Settings"
        ]
      },
      "get": {
        "description": "Retrieve settings for a specific category",
        "operationId": "SettingsController_getCategory",
        "parameters": [
          {
            "name": "category",
            "required": true,
            "in": "path",
            "description": "Settings category (notification, system, communication, profile, emergency)",
            "schema": {
              "example": "notification",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category settings retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get specific settings category",
        "tags": [
          "Settings"
        ]
      }
    }
  },
  "info": {
    "title": "Emergency Project API",
    "description": "\n      ## Emergency Response Management System API\n      \n      This API provides comprehensive endpoints for managing emergency requests, hospitals, rescue teams, and related services.\n      \n      ### Features:\n      - **Authentication**: OAuth 2.0 (Google, Facebook, Apple) and Email/Password\n      - **Emergency Requests**: Create and manage SOS emergency requests\n      - **Hospital Management**: Hospital capacity, emergency acceptance\n      - **Rescue Teams**: Team management and assignment\n      - **Dashboard**: Real-time statistics and monitoring\n      - **Notifications**: Real-time push notifications\n      - **Reports**: Generate and download reports\n      - **User Settings**: Manage user preferences and settings\n      \n      ### Authentication:\n      Most endpoints require JWT Bearer token authentication. Use the /auth endpoints to obtain tokens.\n    ",
    "version": "1.0",
    "contact": {
      "name": "Emergency Project Team",
      "url": "https://emergency-project.com",
      "email": "support@emergency-project.com"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local Development Server"
    },
    {
      "url": "https://api.emergency-project.com",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateEmergencyRequestDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "Description of the emergency"
          },
          "location": {
            "type": "string",
            "description": "Location address"
          },
          "latitude": {
            "type": "number",
            "description": "Latitude of the location"
          },
          "longitude": {
            "type": "number",
            "description": "Longitude of the location"
          },
          "grade": {
            "type": "string",
            "enum": [
              "CRITICAL",
              "URGENT",
              "NON_URGENT"
            ],
            "description": "Grade of the emergency"
          },
          "type": {
            "type": "string",
            "enum": [
              "ACCIDENT",
              "MEDICAL",
              "FIRE",
              "CRIME",
              "OTHER"
            ],
            "description": "Type of the emergency"
          },
          "medicalInfo": {
            "type": "object",
            "description": "Additional medical information"
          },
          "symptoms": {
            "description": "Symptoms reported",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "patientId": {
            "type": "string",
            "description": "Patient ID (optional for EMERGENCY_CENTER)"
          }
        },
        "required": [
          "description",
          "grade",
          "type"
        ]
      },
      "UpdateEmergencyStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ASSIGNED",
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELLED"
            ],
            "description": "New status of the emergency"
          },
          "notes": {
            "type": "string",
            "description": "Additional notes"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "UpdateSettingsDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}